Add the ability to insert config sections directly (#333)
Without this ability extra config secrets need to be configured
manually and in an all-helm environment that means an extra chart just
to add a secret. With this addition the config can be put in a local
values file and be used directly by kubed along with the other config
here.